+++ gp2gp-translator/src/main/java/uk/nhs/adaptors/pss/translator/task/EhrExtractMessageHandler.java
@@ -232,7 +232,8 @@ public class EhrExtractMessageHandler {
             conversationId,
             patientNhsNumber,
             migrationRequest.getWinningPracticeOdsCode(),
-            migrationStatusLog.getDate().toInstant()
+            migrationStatusLog.getDate().toInstant(),
+            messageId
         );
     }
 
@@ -305,10 +306,18 @@ public class EhrExtractMessageHandler {
         String conversationId,
         String patientNhsNumber,
         String winningPracticeOdsCode,
-        Instant mcciIN010000UK13creationTime
+        Instant mcciIN010000UK13creationTime,
+        String ehrExtractId
     ) {
         sendContinueRequestHandler.prepareAndSendRequest(
-            prepareContinueRequestData(payload, conversationId, patientNhsNumber, winningPracticeOdsCode, mcciIN010000UK13creationTime)
+            prepareContinueRequestData(
+                payload,
+                conversationId,
+                patientNhsNumber,
+                winningPracticeOdsCode,
+                mcciIN010000UK13creationTime,
+                ehrExtractId
+            )
         );
     }
 
@@ -317,7 +326,8 @@ public class EhrExtractMessageHandler {
         String conversationId,
         String patientNhsNumber,
         String winningPracticeOdsCode,
-        Instant mcciIN010000UK13creationTime
+        Instant mcciIN010000UK13creationTime,
+        String ehrExtractId
     ) {
         var fromAsid = XmlParseUtilService.parseFromAsid(payload);
         var toAsid = XmlParseUtilService.parseToAsid(payload);
@@ -332,6 +342,7 @@ public class EhrExtractMessageHandler {
             .toOdsCode(toOdsCode)
             .fromOdsCode(winningPracticeOdsCode)
             .mcciIN010000UK13creationTime(mcciIN010000UK13creationTimeToHl7Format)
+            .ehrExtractId(ehrExtractId)
             .build();
     }